+++ app/about/page.tsx
@@ -1,8 +1,8 @@
 import path from 'path';
 import { Fragment, Suspense } from 'react';
-import GithubContributions from '../../components/GithubContributions/GithubContributions';
 import TopTracks from '../../components/Spotify/TopTracks';
 import { readMDXFile } from '../blog/utils';
+import GithubContributions from '../components/github-contributions/github-contributions';
 import { CustomMDX } from '../components/mdx';
 import Occupation from './occupation';
