+++ app/components/command-palette/CommandItem.tsx
