+++ app/components/github-contributions/helpers.ts
