+++ app/page.tsx
@@ -1,8 +1,8 @@
-import Contact from '@/components/Contact/Contact';
 import Hero from '@/components/Hero';
 import Intro from '@/components/Intro';
 import { ScrollProvider } from '@/components/Providers/ScrollProvider';
 import Works from '@/components/Work/Works';
+import Contact from 'app/components/contact';
 
 export default function Page() {
   return (
